# web/src/lib/openapi.schema.json
{
  "components": {
    "schemas": {
...
      "AnalyzeEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ProgressEvent"
          },
          {
            "$ref": "#/components/schemas/ExplainCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/RewritesTestedEvent"
          },
          {
            "$ref": "#/components/schemas/ReadysetCheckedEvent"
          },
          {
            "$ref": "#/components/schemas/CompleteEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      },
      "AnalyzeRequest": {
        "properties": {
          "fast": {
            "default": false,
            "title": "Fast",
            "type": "boolean"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "skip_readyset": {
            "default": false,
            "title": "Skip Readyset",
            "type": "boolean"
          },
          "skip_rewrites": {
            "default": false,
            "title": "Skip Rewrites",
            "type": "boolean"
          },
          "skip_storage": {
            "default": false,
            "title": "Skip Storage",
            "type": "boolean"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "query"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "AskEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/AskStatusEvent"
          },
          {
            "$ref": "#/components/schemas/AskSchemaLoadedEvent"
          },
          {
            "$ref": "#/components/schemas/AskClarificationNeededEvent"
          },
          {
            "$ref": "#/components/schemas/AskSqlGeneratedEvent"
          },
          {
            "$ref": "#/components/schemas/AskResultEvent"
          },
          {
            "$ref": "#/components/schemas/AskErrorEvent"
          }
        ]
      },
...
      "AskRequest": {
        "description": "Request for text-to-SQL conversion.",
        "properties": {
          "agent_mode": {
            "default": false,
            "title": "Agent Mode",
            "type": "boolean"
          },
          "clarification_answers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification Answers"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "selected_interpretation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Interpretation Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "timeout": {
            "default": 30,
            "title": "Timeout",
            "type": "integer"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
...
      "AskStatusEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AskPhase"
              },
              {
                "type": "string"
              }
            ],
            "title": "Phase"
          },
          "type": {
            "const": "status",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "phase",
          "message"
        ],
        "title": "AskStatusEvent",
        "type": "object"
      },
...
      "BenchmarkRequest": {
        "description": "Request to run benchmark on queries.",
        "properties": {
          "concurrency": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 1,
            "title": "Concurrency"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 30,
            "title": "Duration Seconds"
          },
          "interval_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 100,
            "title": "Interval Ms"
          },
          "max_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Count"
          },
          "mode": {
            "default": "interval",
            "enum": [
              "interval",
              "concurrency"
            ],
            "title": "Mode",
            "type": "string"
          },
          "queries": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/BenchmarkQueryInput"
                }
              ]
            },
            "title": "Queries",
            "type": "array"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "queries"
        ],
        "title": "BenchmarkRequest",
        "type": "object"
      },
...
      "ErrorEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "partial_results": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Partial Results"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Stage"
          },
          "type": {
            "const": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
...
      "ExplainCompleteEvent": {
        "properties": {
          "cost_estimate": {
            "title": "Cost Estimate",
            "type": "number"
          },
          "database_engine": {
            "title": "Database Engine",
            "type": "string"
          },
          "execution_time_ms": {
            "title": "Execution Time Ms",
            "type": "number"
          },
          "explain_plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Explain Plan"
          },
          "rows_examined": {
            "title": "Rows Examined",
            "type": "integer"
          },
          "rows_returned": {
            "title": "Rows Returned",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "type": {
            "const": "explain_complete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "success",
          "database_engine",
          "execution_time_ms",
          "rows_examined",
          "rows_returned",
          "cost_estimate"
        ],
        "title": "ExplainCompleteEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProgressEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "percent": {
            "title": "Percent",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "type": {
            "const": "progress",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "stage",
          "percent",
          "message"
        ],
        "title": "ProgressEvent",
        "type": "object"
      },
...
      "QueryBenchmarkEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/QueryBenchmarkProgressEvent"
          },
          {
            "$ref": "#/components/schemas/QueryBenchmarkCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/QueryBenchmarkErrorEvent"
          }
        ]
      },
...
      "RewritesTestedEvent": {
        "properties": {
          "best_rewrite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestedRewrite"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Message"
          },
          "original_performance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RewritePerformance"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "rewrite_results": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TestedRewrite"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Rewrite Results"
          },
          "skipped_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Skipped Reason"
          },
          "tested": {
            "title": "Tested",
            "type": "boolean"
          },
          "type": {
            "const": "rewrites_tested",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "tested"
        ],
        "title": "RewritesTestedEvent",
        "type": "object"
      },
...
      "ScanEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ScanStatusEvent"
          },
          {
            "$ref": "#/components/schemas/ScanFilesFoundEvent"
          },
          {
            "$ref": "#/components/schemas/ScanProgressEvent"
          },
          {
            "$ref": "#/components/schemas/ScanQueryResultEvent"
          },
          {
            "$ref": "#/components/schemas/ScanRegistryEvent"
          },
          {
            "$ref": "#/components/schemas/ScanCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/ScanErrorEvent"
          }
        ]
      },
...
      "ScanRequest": {
        "description": "Request body for scan endpoint.",
        "properties": {
          "analyze": {
            "default": false,
            "title": "Analyze",
            "type": "boolean"
          },
          "check": {
            "default": false,
            "title": "Check",
            "type": "boolean"
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff"
          },
          "directory": {
            "title": "Directory",
            "type": "string"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "fail_threshold": {
            "default": 40,
            "title": "Fail Threshold",
            "type": "integer"
          },
          "file_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Pattern"
          },
          "nosave": {
            "default": false,
            "title": "Nosave",
            "type": "boolean"
          },
          "shallow": {
            "default": false,
            "title": "Shallow",
            "type": "boolean"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "warn_threshold": {
            "default": 60,
            "title": "Warn Threshold",
            "type": "integer"
          }
        },
        "required": [
          "target",
          "directory"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
...
      "TopEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TopStatusEvent"
          },
          {
            "$ref": "#/components/schemas/TopConnectedEvent"
          },
          {
            "$ref": "#/components/schemas/TopSourceFallbackEvent"
          },
          {
            "$ref": "#/components/schemas/TopDbLimitWarningEvent"
          },
          {
            "$ref": "#/components/schemas/TopQueriesEvent"
          },
          {
            "$ref": "#/components/schemas/TopQuerySavedEvent"
          },
          {
            "$ref": "#/components/schemas/TopCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/TopErrorEvent"
          }
        ]
      },
...
    }
  },
  "info": {
    "description": "API server for RDST web client",
    "title": "RDST API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze": {
      "post": {
        "operationId": "analyze_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze"
      }
    },
    "/api/analyze/quick": {
      "post": {
        "operationId": "analyze_quick_api_analyze_quick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Quick"
      }
    },
    "/api/ask": {
      "post": {
        "operationId": "ask_api_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/AskEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "ask"
        ]
      }
    },
...
    "/api/query-registry/benchmark": {
      "post": {
        "description": "Run benchmark on queries with live progress updates via SSE.\n\nReturns Server-Sent Events with progress updates during execution.",
        "operationId": "run_benchmark_api_query_registry_benchmark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/QueryBenchmarkEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Benchmark"
      }
    },
...
    "/api/scan": {
      "post": {
        "description": "Scan a directory for ORM queries via SSE streaming.\n\nStreams real-time progress as the scan progresses through phases:\n- `status`: Phase status messages\n- `files_found`: Files with ORM patterns discovered\n- `progress`: Per-phase progress (current/total)\n- `query_result`: Individual query results\n- `registry`: Registry save results\n- `complete`: Scan complete with summary\n- `error`: Error occurred",
        "operationId": "scan_directory_api_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ScanEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Directory",
        "tags": [
          "scan"
        ]
      }
    },
...
    "/api/top/realtime": {
      "get": {
        "description": "Realtime top queries via SSE streaming.\n\nStreams active query monitoring data from pg_stat_activity (PostgreSQL)\nor SHOW FULL PROCESSLIST (MySQL) with 200ms polling.\n\nEvents:\n- `status`: Progress messages\n- `connected`: Connection established (target, engine, source)\n- `queries`: Current top queries (repeats every poll)\n- `query_saved`: New query saved to registry\n- `complete`: Streaming complete (only if duration specified)\n- `error`: Error occurred\n\nExample:\n```javascript\nconst eventSource = new EventSource('/api/top/realtime?target=mydb&duration=30');\n\neventSource.addEventListener('queries', (event) => {\n    const data = JSON.parse(event.data);\n    console.log('Top queries:', data.queries);\n});\n\neventSource.addEventListener('complete', (event) => {\n    eventSource.close();\n});\n```",
        "operationId": "get_top_queries_realtime_api_top_realtime_get",
        "parameters": [
          {
            "description": "Number of top queries to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Number of top queries to return",
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Duration in seconds (auto-complete after N seconds)",
            "in": "query",
            "name": "duration",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Duration in seconds (auto-complete after N seconds)",
              "title": "Duration"
            }
          },
          {
            "description": "Auto-save queries to registry",
            "in": "query",
            "name": "auto_save",
            "required": false,
            "schema": {
              "default": true,
              "description": "Auto-save queries to registry",
              "title": "Auto Save",
              "type": "boolean"
            }
          },
          {
            "description": "Minimum executions/frequency",
            "in": "query",
            "name": "min_freq",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Minimum executions/frequency",
              "minimum": 0,
              "title": "Min Freq",
              "type": "integer"
            }
          },
          {
            "description": "Minimum load percentage",
            "in": "query",
            "name": "min_load_pct",
            "required": false,
            "schema": {
              "default": 0.0,
              "description": "Minimum load percentage",
              "minimum": 0.0,
              "title": "Min Load Pct",
              "type": "number"
            }
          },
          {
            "description": "Minimum total time in seconds",
            "in": "query",
            "name": "min_total_time_s",
            "required": false,
            "schema": {
              "default": 0.0,
              "description": "Minimum total time in seconds",
              "minimum": 0.0,
              "title": "Min Total Time S",
              "type": "number"
            }
          },
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TopEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Top Queries Realtime",
        "tags": [
          "top"
        ]
      }
    },
...
  }
}